Handling Data

Atomic structures and corresponding energy, force and virial values are read from .xyz files. Filename and target property fields are specified via data_params()

ACE1pack.data_params — Function

data_params(; kwargs...)` : returns a dictionary containing the complete set of parameters required to read data from an .xyz file. j All parameters are passed as keyword arguments.

Parameters

fname : a "*.xyz" file with atomistic data (mandatory).

energy_key = "energy" : key identifying energy for fitting.

force_key = "forces" : key identifying forces for fitting.

virial_key = "virial" : key identifying virial tensor for fitting.

weight_key = "config_type` : key identifying label for setting the correct weight from weights dictionary.