Solvers

LSQR, RRQR, BRR & ARD solvers that solve the Ax=b problem with some regularisation are available via ACEfit module. solver_params() is the way these are define and selected in ACEpotentials.

ACEpotentials.solver_params — Function

solver_params(; kwargs...) : returns a dictionary containing the complete set of parameters required to construct one of the solvers. All parameters are passed as keyword argument and the kind of parameters required depend on "type".

QR Parameters

'type = "qr"`

LSQR Parameters

type = "lsqr"

damp = 5e-3

atol = 1e-6

colim = 1e8

maxiter = 1e5

verbose = false

RRQR Parameters

type = "rrqr"

rtol = 1e-5

SKLEARN_BRR

type = "sklearn_brr"

n_iter = 300

tol = 1e-3

SKLEARN_ARD

type = "sklearn_ard"

n_iter = 300

tol = 1e-3

threshold_lambda = 1e4

BLR

type = "blr"

verbose = false

Additional regularizers (currently only "laplacian") are given via regularizer_params()

ACEpotentials.regularizer_params — Function

regularizer_params(; type = "laplacian", kwargs...) : returns a dictionary containing the complete set of parameters required to construct one of the solvers. All parameters are passed as keyword argument and the kind of parameters required depend on "type".

LSQR Parameters (default)

type = "laplacian"

rlap_scal = 3.0